backend/requirements.txt
asgiref>=3.5.1
backports.zoneinfo>=0.2.1
Django>=4.0.4
django-appconf>=1.0.5
django-ckeditor>=6.4.0
django-cors-headers>=3.11.0
django-debug-toolbar>=3.3.0
django-environ>=0.8.1
django-imagekit>=4.1.0
django-js-asset>=2.0.0
django-rest-framework>=0.1.0
django-uuslug>=2.0.0
djangorestframework>=3.13.1
pilkit>=2.0
Pillow>=9.1.0
python-slugify>=6.1.2
pytz>=2022.1
six>=1.16.0
sqlparse>=0.4.2
text-unidecode>=1.3
